routes moved from projects to project folder
